requirements.txt

streamlit
pandas
altair
numpy
matplotlib
pillow
prophet
xgboost
statsmodels
scikit-learn

# Optional
# pmdarima
